readme: initial temporary readme update
Update the repository README file to accurately reflect
the intent and nature of this repository. The original
io.js README is preserved as iojs_README.md.

This is temporary just while the convergence process
gets underway. Once the repository moves over to
the official organization, this can be reverted back to
an actual project README that accurately reflects the
combined projects.

Reviewed-By: @jasnell James M Snell <jasnell@gmail.com>
Reviewed-By: @Fishrock123 Jeremiah Senkpiel
PR-URL: https://github.com/jasnell/node.js-convergence/pull/4